requirements.txt
tensorflow
keras
numpy
opencv-python
matplotlib
scikit-learn
streamlit
pillow
